the bot's file structure has been updated
