+++ src/main.py
@@ -1,7 +1,9 @@
 # мэйн - точка входа, в мэйне проиcходит аунтетифакция пользователя, инициализация хэндлеров...
+# пока тут просто шаблон
 
 import logging
-from bot_token import token # if highlited see requriments in README.md
+from bot_token import token  # if highlighted see requirements in README.md
+from modules.general import echo, start, help_command
 
 from telegram import ForceReply, Update
 from telegram.ext import Application, CommandHandler, ContextTypes, MessageHandler, filters
@@ -16,31 +18,11 @@ logging.getLogger("httpx").setLevel(logging.WARNING)
 logger = logging.getLogger(__name__)
 
 
-# Define a few command handlers. These usually take the two arguments update and
-# context.
-async def start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Send a message when the command /start is issued."""
-    user = update.effective_user
-    await update.message.reply_html(
-        rf"Hi {user.mention_html()}!",
-        reply_markup=ForceReply(selective=True),
-    )
-
-
-async def help_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Send a message when the command /help is issued."""
-    await update.message.reply_text("Help!")
-
-
-async def echo(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
-    """Echo the user message."""
-    await update.message.reply_text(update.message.text)
-
-
 def main() -> None:
     """Start the bot."""
     # Create the Application and pass it your bot's token.
     application = Application.builder().token(token).build()
+    
 
     # on different commands - answer in Telegram
     application.add_handler(CommandHandler("start", start))
