+++ src/modules/admin.py
@@ -1,3 +1,4 @@
 # Админ может делать всё, быть барменом, либо заказчиком. 
 # Также какие-то логи возможно смотреть и прочую техническую часть. 
 # Назначать бармена, и прочее прочее.
+# Здеся либо класс, либо просто написать логику для менюшек, какие-то функции только для админа и прочее, прочее.
