+++ src/modules/barman.py
@@ -1,2 +1,3 @@
 # Бармен имеет очередь заказов, он их принимает и отмечает выполненными.
 # Также может заказать что-то сам, либо стать заказчиком.
+# Здеся либо класс, либо просто написать логику для менюшек, какие-то функции только для бармена и прочее, прочее.
