+++ src/modules/customer.py
@@ -1,2 +1,3 @@
 # Покупатель (Заказывающий) может заказать коктейль из списка.
 # При заказе должно быть оконо подтверждения или отмены заказа, после этого заказ попадает в очередь к бармену.
+# Здеся либо класс, либо просто написать логику для менюшек, какие-то функции только для покупателя и прочее, прочее.
